fix build + split files
